Extract I/O helpers from session.py into dsgrid/ibis/io.py
Moves the file-based read/write surface out of session.py:
- read_dataframe, try_read_dataframe
- write_dataframe, write_dataframe_and_auto_partition,
  overwrite_dataframe_file, persist_table, persist_intermediate_query
- _read_with_runtime, _read_natively, _post_process_dataframe,
  _is_duckdb_io_exception, _is_spark_parquet_schema_exception,
  _write_table, _table_count
- CsvPartitionWriter
- MAX_PARTITION_SIZE_MB

The lower-level read_csv / read_json / read_parquet that already lived
in io.py are unchanged. The two helpers that reach back into session
state (_read_natively needs createDataFrame, _write_table needs Spark
session.table()) use lazy local imports of get_runtime_session /
get_spark_session to break the session<->io circular dependency.

session.py shrinks by ~420 lines and now re-imports the symbols its
internal create_dataframe / create_dataframe_from_product helpers still
need (_post_process_dataframe, CsvPartitionWriter, MAX_PARTITION_SIZE_MB).
Also drops now-unused session.py imports (math, shutil, time,
load_data, delete_if_exists, Timer, DSGInvalidFile).

15 call sites updated across dsgrid/ and 3 test files. All tests pass
except the existing _DuckDBReader.csv WIP failures, which are
unrelated to this change.
