+++ dsgrid/registry/filesystem_data_store.py
@@ -6,7 +6,7 @@ from typing import Self
 from dsgrid.registry.data_store_interface import DataStoreInterface
 from dsgrid.utils.files import delete_if_exists
 from dsgrid.ibis.operations import coalesce
-from dsgrid.ibis.session import read_dataframe, write_dataframe, write_dataframe_and_auto_partition
+from dsgrid.ibis.io import read_dataframe, write_dataframe, write_dataframe_and_auto_partition
 
 
 TABLE_FILENAME = "table.parquet"
