+++ tests/test_file_schema.py
@@ -26,7 +26,7 @@ from dsgrid.ibis.session import (
     get_runtime_session,
     use_duckdb,
 )
-from dsgrid.ibis.session import write_dataframe
+from dsgrid.ibis.io import write_dataframe
 from dsgrid.ibis.tz import set_session_time_zone
 
 from tests._helpers import collect as _collect, count as _count, order_by as _order_by
